pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solarbalance"
version = "1.0.0"
description = "Астрологический Telegram-бот с точными расчетами и ИИ-прогнозами"
authors = [
    {name = "SolarBalance Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.10.11,<3.13",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "pyswisseph>=2.10.3.2",
    "python-dotenv>=1.0.0",
    "geopy>=2.4.1",
    "timezonefinder>=6.5.0",
    "openai>=1.12.0",
    "skyfield>=1.48",
    "matplotlib>=3.8.2",
    "pillow>=10.2.0",
    "numpy>=1.26.3",
    "scipy>=1.12.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

prod = [
    "asyncpg>=0.28.0",
    "uvloop>=0.18.0",
    "gunicorn>=21.2.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["handlers", "services", "models", "database"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["handlers", "services", "models", "database"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["pyephem.*", "bothub_client.*"]
ignore_missing_imports = true 